Change default settings in GCHP benchmark run directory to match GCC
The GCHP benchmark run directory created with Run/createRunDir.sh now
is automatically set to run a 1-month simulation at c48 grid resolution
for July 2016. Diagnostics are set to be 774-hr (1-month) average
except where explicitly labeled as instantaneous. A custom benchmark
run script for use on Odyssey is copied to the run directory automatically
for immediate use.

The procedure to run a GCHP benchmark run is now 4 steps as follows:
1. Create benchmark run directory from source code
2. Set environment script
  (./setEnvironment environmentFileSamples/gchp.ifort17_openmpi_odyssey.env)
3. make build
4. sbatch gchp.run

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>