Update expected GCHP.rc grid strings in offline advection gridcomp file
This commit corrects a bug in AdvCore_GridCompMod.F90 that resulted in
offsets in the cubed sphere grid areas per face. The problem was that
AdvCore_GridCompMod.F90’s expected strings for the grid parameters in
the config file (GCHP.rc for GCHP) were not consistent with the new
grid parameter strings in MAPL. Both MAPL_Generic.F90 and
MAPL_ExtDataGridCompMod.F90 now use grid config parameter string
that include CF component name and separator ‘.’. For example, instead
of ‘IM:’ in GCHP.rc, MAPL now expected ‘GCHP.IM:’. However,
AdvCore_GridCompMod.F90 still expected ‘IM:’ etc in the file.

If IM and JM are not found then FV_Atm(1)%flagstruct%npx is given a
default value of 32, and that ends up being the value of both
FV_Atm(1)%bd%iec and FV_Atm(1)%bd%jec which are used as end indexes
to extract the area. If running with c24, since the start indexes are 1,
the 1:32 subsetting per face made the area look offset by 1/3.


Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>