Skip non-advected species MBO and C2H4 in GCHP SpeciesConc collection
This is a temporary fix since GEOS-Chem throws an error that these
species are not found in the registry. These species are hard-coded in
the GCHP HISTORY.rc but are not in the GEOS-Chem HISTORY.rc (wildcards
are used instead). This issue will remain open to be looked at in the
future.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>