Comment out emitted Br2 in offline seasalt to match GCC online settings
Br2 is disabled in the GEOS-Chem Classic online seasalt settings with
this line:

--> Emit Br2               :       false

However, Br2 was included as an emitted species in the offline seasalt
emissions used in GCHP. This resulted in differences in the GCHP versus
GEOS-Chem Classic comparisons. Seasalt emissions of Br2 is now disabled
by default in GCHP to match GEOS-Chem Classic.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>