Add commented out debug code to print ESMF field name added by MAPL
This code may be uncommented to view all fields added to an ESMF
state, a helpful feature to debugging problems with GCHP I/O.

This includes all fields in HISTORY.rc that are part of
collections that are turned on (EXPORT), all internal state
variables (INTERNAL), and all fields read via ExtData (IMPORT).
This debugging code will be replaced with debug code that can
be enabled with a run-time parameter once we upgrade to a newer
version of MAPL. Until then this code must be uncommented and
MAPL recompiled for use.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>