Enable saving out fertilizer NOx emissions in HEMCO_Diagn.rc
It turns out to save out manual diagnostics in HEMCO, you simply need to
use the same name in HEMCO_Diagn.rc that is used when creating and updating
the diagnostic within the source code. The fertilizer NOx diagnostic was
formerly named 'FERTILIZER_NO' in the code, but has been changed to
'EmisNO_Fert' for consistency with the naming convention currently used in
HEMCO_Diagn.rc.

PLEASE NOTE: EmisNO_Soil includes fertilizer NOx.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>