Update sample run scripts to log to gchp.log file; remove runConfig.log
runConfig.sh settings and output generated by sourcing the environment
file are now include in gchp.log alongside the run stdout. Consolidating
the information to one location will make debugging and record-keeping
easier.

For the case of the multi-run option, the output of the first sourcing of
runConfig.sh is sent to multirun.log. Subsequent outputs continue to be
sent to the SLURM stdout file in the example multi-run script.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>