Add build_all and clean_all options; fix minor bugs; update make help
build_all and clean_all do the same as build_esmf and clean_esmf but
are more intuitive. Once ESMF is removed from the GCHP repository these
build commands will not need to change other than their definition within
the Makefile (will then be updated to be the same as build_mapl and
clean_mapl).

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>