Fix MAPL ExtData bug: xTSeries allocated without deallocating previous
This update corrects the ExtData problem previously encountered in new
MAPL where the program would crash without an error after extrapolating
year. The problem was that the xTSeries array was allocated within the
extrapolation of year block, and then allocated again later. A check
to deallocate first if allocated was missing.

NOTE: to enable the allow extrapolation option ExtData.rc must have
Ext_AllowExtrap: .true.

The TransportTracers simulations runs to completion out-of-the-box with
this commit.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>