Update Makefile to remove geos executable when doing 'make build_core'
When doing 'make build_core', the geos executable from a previous compilation
may still remain in your run directory and/or code directory. If you
encounter an error in the compilation, build.sh still sees the executable
and returns the message "###   GCHP executable exists!  ###" which can
be misleading. The build_core option now removes the geos executable from
the code and run directories first so that if a compilation fails the user
will see the message "###   WARNING: GCHP executable does not exist   ###'.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>