Fix bug in NEI99 season config entries
This commit fixes a bug in GCHP ExtData.rc that only triggers an error if
starting a run on a day other than the first of the month.
NEI99.season.geos.1x1.nc is a climatology file with 12 values per
variable, one value per month. The climatology column of ExtData.rc
should therefore be Y, and a monthly temporal frequency should also
be set. In GCHP 12.3 and prior, however, climatology is set to N and
no frequency is given. This makes MAPL look for the current date in
the file and errors when none is found. This problem is only encountered
if not starting on the first of the month since the times in the file
correspond to day 1 across all months.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>