Reimplement day-of-week scale factor file handling in ExtData
This update was originally implemented by Sebastian Eastham in an older
version of MAPL, in GCHP commit 1e459ba5 on June 11, 2017.

NOTE: this code is functional but day-of-week scale factors are not
yet usable in this version of GCHP due to netcdf issues in GMAO_pFIO
that need to be investigated.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>