Fix bug in Transport Tracers ExtData.rc; add related MAPL error handling
"needs information message" is now implemented for the case of
fractional regridding string in ExtData.rc missing semi-colon. The
message includes the import name which directs to user to exactly the
line in ExtData.rc that is problematic. In this case, the issue was that
the climatology column was missing, resulting in the 'F' in the time
frequency being interpreted as fractional regridding. MAPL then exited
with an error since no semi-colon was detected in the regridding entry.

The Transport Tracers simulation now runs to completion with this commit.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>