Add debug print in ExtData Initialize_ primary loop loop to specify var
This debug print is only enabled if DEBUG_LEVEL > 0. It prints out the
variable name at the start of the loop in which netcdf data is read from
import files to extract metadata. If an error is encountered during read
it is then more clear what the source of the issue is.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>