Disable use of f2py in MAPL
This update is a temporary kludge. Using f2py can trigger compile errors
that do not impact successful build of GCHP or GCHP run output, but that
are distracting in the compile log file. GMAO will disable f2py in a
future version of MAPL by default.

This update has not been expanded to CMake files.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>