Use updated C2H6_2010 emissions inventory files and turn on by default
This update was partially submitted by Sebastian Eastham who determined
that the lev dimension present in the old files, although not used,
caused a problem with the new version of MAPL.

Note that while the new data path is updated in both ExtData.rc and
HEMCO_Config.rc, only the entry in ExtData.rc is used. It is only updated
in HEMCO_Config.rc for consistency and easy comparison with settings in
GEOS-Chem Classic.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>