Update Strat Bry data entry in HEMCO_Config.rc to match GC Classic
This update will have no affect on GCHP output since the change
originates in fixing a bug in GEOS-Chem Classic only by making setting
HEMCO to read data on the first of every month. This is already the
read frequency in GCHP set in ExtData.rc. The read frequency set in
HEMCO_Config.rc is ignored by MAPL in GCHP.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>