Update benchmark resource request: change 96 to 48+exclusive
This update also adds code to the benchmark run script to
rename the output restart file. This renaming is needed
in case it is used as input to a new run. Otherwise MAPL
will crash when trying to write out another new restart file
since the file already exists. This issue is specific to the
new version of MAPL going into GCHP 12.5.0.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>