Remove redundant nc extension on checkpoint files; use nc4 for all
Previously the checkpoint files that can be periodically output using
the record frequency option would have extent .nc.{datetime}.nc4. This
is because the baseline name set in GCHP.rc included '.nc'. This update
removes that extension, thereby making the periodic checkpoint not contain
the extraneous '.nc'. It also updates all run scripts to rename the
end-of-run checkpoint to use '.nc4' for consistency with the periodic
checkpoints and the diagnostic files.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>