Turn off tracer mixing ratio scaling with met change at start of timestep
This commit removes logic in the first call to airqnt to turn tracer
mixing ratio scaling on if FV3 is off and it is not the very first
timestep of the run. The scaling is now always off for all cases.

While the previous logic conserved tracer mass with met changes when FV3
was off, it also caused differences in multi-segemented versus single runs
because the scaling was skipped on the very first timestep of the run.
For multi vs single run reproducibility we are therefore removing the
mixing ratio scaling. Now when advection is turned off there are no
introduced differences between multi-segmented and single runs.

IMPORTANT NOTE: users should beware that advection off no longer conserves
tracer mass. This could be corrected by incorporating meteorology into
the GCHP internal state in the future if this functionality is deemed
necessary.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>