Expand multirun error handling to cancel jobs if cap_restart is empty
Previously remaining jobs were only canceled if the date in cap_restart
at the end of a run was the same as the start. However, if an error occurs
during the first run in a sequence then this error would not be triggered
and the next job would start. Catching if cap_restart exists but is empty
catches this case of error.

This commit also increases the default # cores in the multi-run job from
6 to 30 to maximize ExtData speed. Input handling is now parallelize
across cores for one node.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>