GMAO bug fix in FV3 advection
This bug fix was submitted for use in GEOS CTM by Kyle Gerheiser (GMAO)
and is relevant to GCHP. See GEOS-ESM/GEOSgcm_GridComp issue #13:
https://github.com/GEOS-ESM/GEOSgcm_GridComp/issues/13

Without this fix concentrations in GCHP will blow up in advection at
our previous default timesteps. Until the bug was fixed we decreased the
default dynamic timesteps as work-around. Updating the default timesteps
in the run directory creation template will be in a later commit.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>