Replace 4x5 input files with files regridded to higher resolution
When running at high core counts (>600) ESMF regridding will fail on
input 4x5 files because of limits in domain decomposition. A newer version
of MAPL in the pipeline will include handling to avoid this. In the
meantime, our work-around is to input higher resolution files.

Please note that there will still be problems when using a very very high
core count, on the order of thousands of cores, due to input 2x2.5 files.

Also note that the HEMCO_Config file paths are updates for constency
with ExtData but are not actually used. Input location is read from
ExtData.rc alone.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>