Replace NH3/v2014-07/NH3_geos.2x25.nc with file regridded from 4x5
The existing 2x2.5 file delivered with the 4x5 file in 2014 has
significantly different emission totals from the 4x5. Using it in GCHP
when GCC uses the 4x5 file therefore causes differences in NH3. To
remove these differences, and to avoid using a file with potentially
bad data, a new file was created by conservatively regridding the 4x5
file. This will be used in GCHP so that runs with high core counts
(>600) are possible pending a fix to the ESMF domain decomposition limit
handling in MAPL.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>