Add FracOfTimeInTrop diagnostic to HISTORY.rc file for benchmarks
The FracOfTimeInTrop output (equivalent to ND54 bpch diagnostic) is needed
for computing the MCF and CH4 lifetimes in GCPy. This diagnostic has been
added to GEOS-Chem and can be included in any collection. Here we add it to
the StateMet collection, which is also needed for computing MCF and CH4
lifetimes. This output is only needed for benchmarking and hasn't been
added to the HISTORY.rc files for other simulations but users may choose
to add it manually.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>