Remove reference to SimType in Chem_GridCompMod.F90
Chem_GridCompMod.F90 was still reading the simulation type number, which
has been removed from input.geos file. This was used to add all species,
including chemical species, to the internal state for full-chemistry
simulations; otherwise those species would be initialized to zero. A simple
fix is to make sure the number of advected species is above a certain value
that would exclude specialty simulations, as is done for MODEL_GEOS. Here we
have chosen 50 because all full-chemistry simulations currently include more
than 100 advected species. We should look for a better way to do this in
the future though.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>