Make intelgcc a valid ESMF_COMPILER option
We've been using "intel" as our ESMF_COMPILER, but that's designed for
situations where Intel compilers are being used for Fortran, C, and C++.
We are instead using Intel's Fortran compilers but GNU C and C++ compilers.
This is the cause of two issues:
a) Strange warnings in the build log about "unary operators"
b) Can't compile ESMF with "ESMF_BOPT=g" (debug flags)
This is easy to fix by setting ESMF_COMPILER=intelgcc, but the Makefiles
for GCHP and GCC need to be modified to allow for this. This commit
does that job for GCHP.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>