Add handling for input file positive attribute 'up'
This update corrects a bug starting in 12.5.0 in which the vertical
dimension of input 3D files (non-met) with 72 levels were improperly
handled by MAPL. This resulted in using arrays in GEOS-Chem with
incorrect vertical flipping. The bug was introduced due to the upgrade to
a newer version of MAPL which no longer contained handling for the
positive attribute in input files. The new handling applies an additional
vertical flipping if the positive attribute is present and equals 'up'.
If the attribute is missing or not 'up' then the old handling is used.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>